refactor interfaces into types
