+++ modules/client/src/client-common/index.ts
@@ -1,7 +1,7 @@
-export * from "./interfaces/plugin";
-export * from "./interfaces/core";
-export * from "./interfaces/common";
-export * from "./interfaces/context";
+export * from "./types/plugin";
+export * from "./interfaces";
+export * from "./types/common";
+export * from "./types/context";
 export * from "./core";
 export * from "./context";
 export * from "./context-plugin";
@@ -13,3 +13,5 @@ export {
   UNAVAILABLE_PROPOSAL_METADATA,
   UNSUPPORTED_PROPOSAL_METADATA_LINK,
 } from "./constants";
+
+export * from "./types";
