+++ modules/client/src/client-common/types/common.ts
@@ -33,7 +33,7 @@ export enum DaoRole {
  * Contains the payload passed to the global DAO factory so that
  * plugins can be initialized
  */
-export interface IPluginInstallItem {
+export interface PluginInstallItem {
   id: string; // ENS domain or address of the plugin's Repo
   data: Uint8Array;
 }
@@ -77,8 +77,14 @@ export enum SortDirection {
   DESC = "desc",
 }
 
-export interface IInterfaceParams {
+export interface InterfaceParams {
   id: string;
   functionName: string;
   hash: string;
+}
+
+export enum TokenType {
+  NATIVE = "native",
+  ERC20 = "erc20",
+  ERC721 = "erc721",
 }
