+++ modules/client/src/multisig/index.ts
@@ -1,2 +1,3 @@
-export { MultisigClient } from "./client";
+export * from "./client";
+export * from "./types";
 export * from "./interfaces";
