+++ modules/client/src/multisig/internal/constants.ts
@@ -1,8 +1,6 @@
 // @ts-ignore
 // todo fix new contracts-ethers
-import {
-  Multisig__factory,
-} from "@aragon/osx-ethers";
+import { Multisig__factory } from "@aragon/osx-ethers";
 
 // TODO update with function names
 export const AVAILABLE_FUNCTION_SIGNATURES: string[] = [
