+++ modules/client/src/multisig/internal/graphql-queries/index.ts
@@ -1,2 +1,3 @@
 export * from "./proposal";
 export * from "./settings";
+export * from "./members";
