+++ modules/client/src/tokenVoting/index.ts
@@ -1,17 +1,3 @@
-export { TokenVotingClient } from "./client";
-export {
-  DelegateTokensParams,
-  DelegateTokensStep,
-  DelegateTokensStepValue,
-  Erc20TokenDetails,
-  IMintTokenParams,
-  ITokenVotingPluginInstall,
-  TokenVotingMember,
-  TokenVotingPluginPrepareInstallationParams,
-  TokenVotingProposal,
-  TokenVotingProposalListItem,
-  TokenVotingProposalResult,
-  UndelegateTokensStepValue,
-  UnwrapTokensStep,
-  WrapTokensStep,
-} from "./interfaces";
+export * from "./client";
+export * from "./interfaces";
+export * from "./types";
