+++ modules/client/src/tokenVoting/internal/client/estimation.ts
@@ -6,7 +6,7 @@ import {
   ClientCore,
   CreateMajorityVotingProposalParams,
   GasFeeEstimation,
-  IVoteProposalParams,
+  VoteProposalParams,
 } from "../../../client-common";
 import {
   DelegateTokensParams,
@@ -74,12 +74,12 @@ export class TokenVotingClientEstimation extends ClientCore
   /**
    * Estimates the gas fee of casting a vote on a proposal
    *
-   * @param {IVoteProposalParams} params
+   * @param {VoteProposalParams} params
    * @return {*}  {Promise<GasFeeEstimation>}
    * @memberof TokenVotingClientEstimation
    */
   public async voteProposal(
-    params: IVoteProposalParams,
+    params: VoteProposalParams,
   ): Promise<GasFeeEstimation> {
     const signer = this.web3.getConnectedSigner();
     if (!signer) {
